Fixed majority of problems, memory allocation still a problem

+++ include/Potentiometer.h
@@ -8,7 +8,7 @@ class Potentiometer : Input
 {
 private:
   uint8_t _pin;
-  uint8_t _value;
+  uint16_t _value;
   uint16_t _numLeds;
 
   void update();
